Backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
dnspython==1.16.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
motor==3.7.0
passlib==1.7.4
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
